fix: renamed the model from Vivek to Jadhav
It is more common to use the surname of the first author and not the name.
