+++ tools/matlab_py_comparison/plot_dog_py_matlab.py
@@ -1,6 +1,6 @@
 import matplotlib.pyplot as plt
 
-from movement_algorithms.vivek_model import herding_model
+from movement_algorithms.jadhav_model import herding_model
 from utils.utils import load_simulation_results_matlab, extract_initial_conditions
 
 matlab_file = "../../data/hm_1_14_new.mat"
